requirements.txt
automationhat==0.0.4
blinker==1.3
blinkt==0.1.0
Cap1xxx==0.1.3
chardet==2.3.0
colorama==0.3.2
drumhat==0.0.5
envirophat==0.0.6
ExplorerHAT==0.4.2
Flask==0.10.1
flickrapi==2.3
fourletterphat==0.0.2
gpiozero==1.4.0
html5lib==0.999
itsdangerous==0.24
Jinja2==2.7.3
lxkeymap==0.1
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.1.3
mote==0.0.3
motephat==0.0.2
ndg-httpsclient==0.3.2
numpy==1.8.2
oauthlib==2.0.2
phatbeat==0.0.2
pi3d==2.19
pianohat==0.0.5
picamera==1.13
picraft==1.0
pifacecommon==4.2.1
pifacedigitalio==3.1.0
piglow==1.2.4
pigpio==1.35
Pillow==2.6.1
psutil==5.2.2
pyasn1==0.1.7
pygame==1.9.2a0
pygobject==3.14.0
pyinotify==0.9.4
pyOpenSSL==0.13.1
pyserial==2.6
python-apt==0.9.3.12
python-xlib==0.19
PyUserInput==0.1.11
rainbowhat==0.0.2
requests==2.4.3
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
RPi.GPIO==0.6.3
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.0.1
sense-emu==1.0
sense-hat==2.2.0
six==1.10.0
skywriter==0.0.7
smbus==1.1
sn3218==1.2.7
spidev==3.0
touchphat==0.0.1
twython==3.1.2
urllib3==1.9.1
Werkzeug==0.9.6
